Refactor shared web module: restructure package hierarchy, remove unused DTOs, and migrate `AuditResponse` and `PersonNameResponse` to `shared.infrastructure.web.responses`.
