+++ src/main/java/by/iivanov/rpm/iam/user/infrastructure/web/UserSummaryResponse.java
@@ -1,7 +1,7 @@
 package by.iivanov.rpm.iam.user.infrastructure.web;
 
-import by.iivanov.rpm.shared.web.AuditResponse;
-import by.iivanov.rpm.shared.web.PersonNameResponse;
+import by.iivanov.rpm.shared.infrastructure.web.responses.AuditResponse;
+import by.iivanov.rpm.shared.infrastructure.web.responses.PersonNameResponse;
 
 /**
  * A user as shown in the admin user list: a summary projection with identity, status, and audit metadata
