+++ src/main/java/by/iivanov/rpm/shared/infrastructure/web/responses/package-info.java
@@ -1,0 +1,6 @@
+@NullMarked
+@InfrastructureRing
+package by.iivanov.rpm.shared.infrastructure.web.responses;
+
+import org.jmolecules.architecture.onion.simplified.InfrastructureRing;
+import org.jspecify.annotations.NullMarked;
