--- src/main/java/by/iivanov/rpm/shared/web/errors/ErrorConstants.java
@@ -1,18 +1,0 @@
-package by.iivanov.rpm.shared.web.errors;
-
-import java.net.URI;
-
-public final class ErrorConstants {
-
-    public static final String ERR_CONCURRENCY_FAILURE = "error.concurrencyFailure";
-    public static final String ERR_VALIDATION = "error.validation";
-    public static final String PROBLEM_BASE_URL = "https://www.rpm-ddd.my/problem";
-    public static final URI DEFAULT_TYPE = URI.create(PROBLEM_BASE_URL + "/problem-with-message");
-    public static final URI CONSTRAINT_VIOLATION_TYPE = URI.create(PROBLEM_BASE_URL + "/constraint-violation");
-    public static final URI INVALID_PASSWORD_TYPE = URI.create(PROBLEM_BASE_URL + "/invalid-password");
-    public static final URI EMAIL_ALREADY_USED_TYPE = URI.create(PROBLEM_BASE_URL + "/email-already-used");
-    public static final URI LOGIN_ALREADY_USED_TYPE = URI.create(PROBLEM_BASE_URL + "/login-already-used");
-    public static final URI ACCESS_DENIED_TYPE = URI.create(PROBLEM_BASE_URL + "/access-denied");
-
-    private ErrorConstants() {}
-}
