--- src/main/java/by/iivanov/rpm/shared/web/errors/FieldErrorDto.java
@@ -1,6 +1,0 @@
-package by.iivanov.rpm.shared.web.errors;
-
-import org.jspecify.annotations.Nullable;
-
-public record FieldErrorDto(
-        String objectName, String field, @Nullable String message) {}
