--- src/main/java/by/iivanov/rpm/shared/web/errors/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package by.iivanov.rpm.shared.web.errors;
-
-import org.jspecify.annotations.NullMarked;
