--- src/main/java/by/iivanov/rpm/shared/web/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package by.iivanov.rpm.shared.web;
-
-import org.jspecify.annotations.NullMarked;
